fix int8 bug

+++ src/meta/meta.go
@@ -21,7 +21,7 @@ func init() {
 	TYPE_DIC = map[string]string{
 		"int":"int64",
 		"int4":"int64",
-		"bigint":"int64",
+		"int8":"int64",
 		"number":"float64",
 		"text":"string",
 		"char":"string",
